Extract engine thumbnail operation
thumbnail is a compact frame-extraction operation with timestamp clamping and output-path generation. This moves it into engine_thumbnail.py while re-exporting thumbnail from mcp_video.engine for compatibility.

Constraint: Existing callers import thumbnail from mcp_video.engine and expect ThumbnailResult behavior unchanged.

Rejected: Move preview/storyboard with thumbnail | those have separate sampling and output contracts and should be reviewed independently.

Confidence: high

Scope-risk: narrow

Directive: Keep extracting engine media output operations in small focused modules.

Tested: ruff check --fix mcp_video/engine.py mcp_video/engine_thumbnail.py

Tested: /opt/homebrew/bin/python3 thumbnail re-export smoke

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py -k 'thumbnail' tests/test_server.py -k 'thumbnail' tests/test_cli.py -k 'thumbnail' tests/test_e2e.py -k 'thumbnail' -q --tb=short

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short
