+++ mcp_video/engine.py
@@ -33,7 +33,6 @@ from .models import (
     SplitLayout,
     StoryboardResult,
     SubtitleResult,
-    ThumbnailResult,
     Timeline,
     TimelineImageOverlay,
     WaveformResult,
@@ -49,6 +48,7 @@ from .engine_text import add_text as add_text
 from .engine_audio_ops import add_audio as add_audio
 from .engine_resize import resize as resize
 from .engine_speed import speed as speed
+from .engine_thumbnail import thumbnail as thumbnail
 from .engine_runtime_utils import (
     _auto_output as _auto_output,
     _auto_output_dir as _auto_output_dir,
@@ -281,46 +281,6 @@ def convert(
         operation="convert",
         progress=100.0,
         thumbnail_base64=thumb_b64,
-    )
-
-
-def thumbnail(
-    input_path: str,
-    timestamp: float | None = None,
-    output_path: str | None = None,
-) -> ThumbnailResult:
-    """Extract a single frame from a video."""
-    _validate_input(input_path)
-
-    if timestamp is None:
-        # Grab frame at 10% of video duration
-        dur = get_duration(input_path)
-        timestamp = dur * 0.1
-    else:
-        # Clamp to valid range
-        dur = get_duration(input_path)
-        timestamp = min(timestamp, dur * 0.99)
-
-    output = output_path or _auto_output(input_path, f"frame_{timestamp:.1f}s", ext=".jpg")
-
-    _run_ffmpeg(
-        [
-            "-ss",
-            str(timestamp),
-            "-i",
-            input_path,
-            "-vframes",
-            "1",
-            "-q:v",
-            "2",
-            "-y",
-            output,
-        ]
-    )
-
-    return ThumbnailResult(
-        frame_path=output,
-        timestamp=timestamp,
     )
